Seperated jobCategory controller and routes from job controller and routes
